Optional TOML Support (disabled by default)
This patch makes TOML support optional, disabling it by default. To
enable support include the Golang build tag `toml`. For example, the
following command builds Viper without TOML support:

        $ go build

Whereas this command will build Viper, along with support for parsing
TOML:

        $ go build -tags toml

The reason for making TOML optional is due to the dependency upon the
BurntSushi package that is used to unmarshal TOML content. The package
is licensed under the WTFPL license (http://www.wtfpl.net/) and
incompatible with the Kubernetes project. Pull requests submitted to K8
that transitively include a dependency upon packages licensed under the
WTFPL license are denied.

To this end, until such time an alternative package is deemed acceptable
for parsing TOML content, this patch would make TOML an optional
component of Viper, enabled only when explicitly requested via the build
tag `toml`.
